+++ src/main/java/com/evacipated/cardcrawl/mod/stslib/Keyword.java
@@ -1,9 +1,12 @@
 package com.evacipated.cardcrawl.mod.stslib;
 
+import basemod.helpers.KeywordColorInfo;
+
 public class Keyword
 {
     public String ID = "";
     public String PROPER_NAME;
     public String[] NAMES;
     public String DESCRIPTION;
+    public KeywordColorInfo COLOR;
 }